pyproject.toml
[project]
name = "nanumsa-server"
version = "0.1.0"
description = ""
authors = [
    {name = "smilemeeso",email = "m950827@naver.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "boto3 (>=1.36.11,<2.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "geoalchemy2[shapely] (>=0.17.0,<0.18.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "korean-name-generator (>=1.0.1,<2.0.0)",
    "datetime (>=5.5,<6.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "aws (>=0.2.5,<0.3.0)",
    "firebase-admin (>=6.6.0,<7.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-socketio[client] (>=5.12.1,<6.0.0)",
    "websocket-client (>=1.8.0,<2.0.0)"
]

[tool.poetry]
packages = [
    { include = "src" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
